gym/atari_breakout/requirements.txt
absl-py==1.4.0
ale-py==0.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
astunparse==1.6.3
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
boto3==1.26.129
botocore==1.29.129
cachetools==5.3.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
contourpy==1.0.7
cycler==0.11.0
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dm-tree==0.1.8
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
Farama-Notifications==0.0.4
filelock @ file:///C:/b/abs_c7yrhs9uz2/croot/filelock_1672387617533/work
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
gymnasium==0.28.1
h5py==3.8.0
idna==3.4
imageio==2.28.0
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1672612343532/work
importlib-resources==5.12.0
ipykernel @ file:///D:/bld/ipykernel_1655369313836/work
ipython @ file:///D:/bld/ipython_1676047671725/work
jax-jumpy==1.0.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
jmespath==1.0.1
joblib==1.2.0
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1676579893731/work
jupyter_core @ file:///D:/bld/jupyter_core_1675110009198/work
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.2
libclang==15.0.6.1
libtorrent==2.0.7
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mpmath==1.2.1
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1673482170163/work
pandas==1.5.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1675735718929/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1670414775770/work
protobuf==3.19.6
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pygame==2.2.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1672682006896/work
pyparsing==3.0.9
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==1.0.0
pytz==2022.7.1
PyWavelets==1.4.1
pywin32==227
pyzmq @ file:///C:/ci/pyzmq_1657615952984/work
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.9.1
Shimmy==0.2.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy @ file:///C:/b/abs_95fbf1z7n6/croot/sympy_1668202411612/work
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.10.0
tensorflow-directml-plugin==0.4.0.dev230202
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.19.0
tensorflow_intel==2.10.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.1
tornado @ file:///D:/bld/tornado_1656937934674/work
tqdm==4.64.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.15.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
